renamed diagnostics to connections for clarity.
